requirements.txt
accelerate==0.24.1
bitsandbytes==0.41.2.post2
blobfile==2.0.2
datasets==2.15.0
ddpm==1.0.0
diffusers==0.15.0
huggingface_hub==0.19.4
mpi4py==3.1.5
numpy==1.24.3
opencv_contrib_python==4.8.1.78
opencv_python==4.6.0.66
opencv_python_headless==4.8.0.76
packaging==21.3
Pillow==10.2.0
PyYAML==6.0.1
tensorflow==2.11.0
torch==1.13.1
torchdiffeq==0.2.3
torchsde==0.2.5
torchvision==0.14.1
tqdm==4.64.1
transformers==4.36.2
wandb==0.12.21
xformers==0.0.17.dev465
